# backend/docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "AI 内容创作平台后端 API。包含用户注册登录、文章多智能体生成、SSE 进度推送、配额管理等。",
        "title": "AI Content Creator API",
        "termsOfService": "https://github.com/ai-content-creator",
        "contact": {
            "name": "API Support",
            "url": "https://github.com/ai-content-creator"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8123",
    "basePath": "/api",
    "paths": {
...
        "/user/add": {
            "post": {
                "security": [
                    {
                        "SessionCookie": []
                    }
                ],
                "description": "管理员创建用户，需 admin 角色",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户管理"
                ],
                "summary": "创建用户",
                "parameters": [
                    {
                        "description": "创建参数",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.AddUserRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "业务错误（无权限、参数错误等）",
                        "schema": {
                            "$ref": "#/definitions/common.BaseResponse"
                        }
                    }
                }
            }
        },
        "/user/delete": {
            "post": {
                "security": [
                    {
                        "SessionCookie": []
                    }
                ],
                "description": "管理员逻辑删除用户，需 admin 角色",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户管理"
                ],
                "summary": "删除用户",
                "parameters": [
                    {
                        "description": "删除参数",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.DeleteRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "业务错误",
                        "schema": {
                            "$ref": "#/definitions/common.BaseResponse"
                        }
                    }
                }
            }
        },
        "/user/get": {
            "get": {
                "security": [
                    {
                        "SessionCookie": []
                    }
                ],
                "description": "管理员获取用户完整信息（含密码哈希等敏感字段），需 admin 角色",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户管理"
                ],
                "summary": "根据 ID 获取用户",
                "parameters": [
                    {
                        "type": "integer",
                        "format": "int64",
                        "description": "用户 ID",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "业务错误",
                        "schema": {
                            "$ref": "#/definitions/common.BaseResponse"
                        }
                    }
                }
            }
        },
        "/user/get/login": {
            "get": {
                "security": [
                    {
                        "SessionCookie": []
                    }
                ],
                "description": "从 Session 读取当前登录用户信息，含配额等",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户"
                ],
                "summary": "获取当前登录用户",
                "responses": {
                    "200": {
                        "description": "未登录",
                        "schema": {
                            "$ref": "#/definitions/common.BaseResponse"
                        }
                    }
                }
            }
        },
        "/user/get/vo": {
            "get": {
                "description": "获取用户公开信息（UserInfo），不含密码",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户"
                ],
                "summary": "根据 ID 获取用户脱敏信息",
                "parameters": [
                    {
                        "type": "integer",
                        "format": "int64",
                        "description": "用户 ID",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "业务错误",
                        "schema": {
                            "$ref": "#/definitions/common.BaseResponse"
                        }
                    }
                }
            }
        },
        "/user/list/page/vo": {
            "post": {
                "security": [
                    {
                        "SessionCookie": []
                    }
                ],
                "description": "管理员分页查询用户，支持多条件筛选与排序，需 admin 角色",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户管理"
                ],
                "summary": "分页查询用户列表",
                "parameters": [
                    {
                        "description": "查询参数",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.QueryUserRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "业务错误",
                        "schema": {
                            "$ref": "#/definitions/common.BaseResponse"
                        }
                    }
                }
            }
        },
        "/user/login": {
            "post": {
                "description": "登录成功后写入 Session Cookie（session），后续受保护接口需携带",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户"
                ],
                "summary": "用户登录",
                "parameters": [
                    {
                        "description": "登录参数",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.LoginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "业务错误（账号或密码错误等）",
                        "schema": {
                            "$ref": "#/definitions/common.BaseResponse"
                        }
                    }
                }
            }
        },
        "/user/logout": {
            "post": {
                "security": [
                    {
                        "SessionCookie": []
                    }
                ],
                "description": "清除 Session，退出登录",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户"
                ],
                "summary": "用户注销",
                "responses": {
                    "200": {
                        "description": "业务错误",
                        "schema": {
                            "$ref": "#/definitions/common.BaseResponse"
                        }
                    }
                }
            }
        },
        "/user/register": {
            "post": {
                "description": "注册新用户，账号至少 4 位，密码至少 8 位",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户"
                ],
                "summary": "用户注册",
                "parameters": [
                    {
                        "description": "注册参数",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.RegisterRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "业务错误（参数错误、账号重复等）",
                        "schema": {
                            "$ref": "#/definitions/common.BaseResponse"
                        }
                    }
                }
            }
        },
        "/user/update": {
            "post": {
                "security": [
                    {
                        "SessionCookie": []
                    }
                ],
                "description": "管理员更新用户信息，需 admin 角色",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户管理"
                ],
                "summary": "更新用户",
                "parameters": [
                    {
                        "description": "更新参数",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.UpdateUserRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "业务错误",
                        "schema": {
                            "$ref": "#/definitions/common.BaseResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "common.BaseResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "业务状态码，0 表示成功，非 0 表示失败",
                    "type": "integer",
                    "example": 0
                },
                "data": {
                    "description": "响应数据，成功时返回；失败时通常为 null"
                },
                "message": {
                    "description": "响应消息，成功时为 ok，失败时为错误描述",
                    "type": "string",
                    "example": "ok"
                }
            }
        },
        "model.AddUserRequest": {
            "type": "object",
            "required": [
                "userAccount"
            ],
            "properties": {
                "userAccount": {
                    "description": "登录账号",
                    "type": "string",
                    "example": "newuser"
                },
                "userAvatar": {
                    "description": "头像 URL（可选）",
                    "type": "string",
                    "example": "https://example.com/avatar.png"
                },
                "userName": {
                    "description": "用户昵称（可选）",
                    "type": "string",
                    "example": "新用户"
                },
                "userProfile": {
                    "description": "个人简介（可选）",
                    "type": "string",
                    "example": "个人简介"
                },
                "userRole": {
                    "description": "用户角色：user / admin / vip",
                    "type": "string",
                    "enum": [
                        "user",
                        "admin",
                        "vip"
                    ],
                    "example": "user"
                }
            }
        },
...
        "model.DeleteRequest": {
            "type": "object",
            "required": [
                "id"
            ],
            "properties": {
                "id": {
                    "description": "要删除的记录主键 ID（用户或文章）",
                    "type": "integer",
                    "example": 1
                }
            }
        },
...
        "model.LoginRequest": {
            "type": "object",
            "required": [
                "userAccount",
                "userPassword"
            ],
            "properties": {
                "userAccount": {
                    "description": "登录账号",
                    "type": "string",
                    "minLength": 4,
                    "example": "zhangsan"
                },
                "userPassword": {
                    "description": "登录密码",
                    "type": "string",
                    "minLength": 8,
                    "example": "12345678"
                }
            }
        },
...
        "model.QueryUserRequest": {
            "type": "object",
            "properties": {
                "id": {
                    "description": "按用户 ID 筛选（可选）",
                    "type": "integer",
                    "example": 1
                },
                "pageNum": {
                    "description": "页码，从 1 开始",
                    "type": "integer",
                    "example": 1
                },
                "pageSize": {
                    "description": "每页条数",
                    "type": "integer",
                    "example": 10
                },
                "sortField": {
                    "description": "排序字段（可选）",
                    "type": "string",
                    "example": "createTime"
                },
                "sortOrder": {
                    "description": "排序方向：asc / desc（可选）",
                    "type": "string",
                    "example": "desc"
                },
                "userAccount": {
                    "description": "按账号模糊筛选（可选）",
                    "type": "string",
                    "example": "zhangsan"
                },
                "userName": {
                    "description": "按昵称模糊筛选（可选）",
                    "type": "string",
                    "example": "张三"
                },
                "userProfile": {
                    "description": "按简介模糊筛选（可选）",
                    "type": "string"
                },
                "userRole": {
                    "description": "按角色筛选（可选）",
                    "type": "string",
                    "example": "user"
                }
            }
        },
        "model.RegisterRequest": {
            "type": "object",
            "required": [
                "checkPassword",
                "userAccount",
                "userPassword"
            ],
            "properties": {
                "checkPassword": {
                    "description": "确认密码，需与 userPassword 一致",
                    "type": "string",
                    "minLength": 8,
                    "example": "12345678"
                },
                "userAccount": {
                    "description": "登录账号，至少 4 位",
                    "type": "string",
                    "minLength": 4,
                    "example": "zhangsan"
                },
                "userPassword": {
                    "description": "登录密码，至少 8 位",
                    "type": "string",
                    "minLength": 8,
                    "example": "12345678"
                }
            }
        },
...
        "model.UpdateUserRequest": {
            "type": "object",
            "required": [
                "id"
            ],
            "properties": {
                "id": {
                    "description": "用户主键 ID",
                    "type": "integer",
                    "example": 1
                },
                "userAvatar": {
                    "description": "头像 URL（可选）",
                    "type": "string",
                    "example": "https://example.com/avatar.png"
                },
                "userName": {
                    "description": "用户昵称（可选，传 null 表示不更新）",
                    "type": "string",
                    "example": "新昵称"
                },
                "userProfile": {
                    "description": "个人简介（可选）",
                    "type": "string",
                    "example": "更新后的简介"
                },
                "userRole": {
                    "description": "用户角色（可选）",
                    "type": "string",
                    "enum": [
                        "user",
                        "admin",
                        "vip"
                    ],
                    "example": "vip"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "SessionCookie": {
            "description": "登录成功后服务端写入 Session Cookie，受保护接口需携带",
            "type": "apiKey",
            "name": "session",
            "in": "cookie"
        }
    }
}